requirements.txt
absl-py==2.1.0
aiohappyeyeballs==2.4.3
aiohttp==3.11.7
aiosignal==1.3.1
asttokens==2.4.1
astunparse==1.6.3
async-timeout==5.0.1
attrs==24.2.0
certifi==2024.8.30
charset-normalizer==3.4.0
chex==0.1.87
coloredlogs==15.0.1
comm==0.2.2
contourpy==1.3.1
cycler==0.12.1
debugpy==1.8.8
decorator==5.1.1
etils==1.10.0
exceptiongroup==1.2.2
executing==2.1.0
filelock==3.13.1
flatbuffers==24.3.25
flax==0.10.2
fonttools==4.55.0
frozenlist==1.5.0
fsspec==2024.2.0
gast==0.6.0
google-pasta==0.2.0
grpcio==1.68.0
h5py==3.12.1
humanfriendly==10.0
humanize==4.11.0
idna==3.10
importlib_resources==6.4.5
ipykernel==6.29.5
ipython==8.29.0
jax==0.4.35
jaxlib==0.4.35
jedi==0.19.2
Jinja2==3.1.3
jupyter_client==8.6.3
jupyter_core==5.7.2
keras==3.6.0
kiwisolver==1.4.7
libclang==18.1.1
lightning-utilities==0.11.9
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.2
matplotlib-inline==0.1.7
mdurl==0.1.2
ml-dtypes==0.4.1
mpmath==1.3.0
msgpack==1.1.0
multidict==6.1.0
namex==0.0.8
nest-asyncio==1.6.0
networkx==3.2.1
numpy==1.26.4
onnx==1.17.0
onnx-tf==1.10.0
onnxruntime==1.20.1
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
opt_einsum==3.4.0
optax==0.2.4
optree==0.13.1
orbax-checkpoint==0.10.1
packaging==23.2
pandas==2.2.3
parso==0.8.4
pexpect==4.9.0
pillow==10.2.0
platformdirs==4.3.6
prompt_toolkit==3.0.48
propcache==0.2.0
protobuf==5.28.3
psutil==6.1.0
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.18.0
pyparsing==3.2.0
python-dateutil==2.9.0.post0
pytorch-lightning==2.4.0
pytorch-triton-rocm @ file:///home/abendafi/gan_model/pytorch_triton_rocm-2.1.0%2Brocm6.1.3.4d510c3a44-cp310-cp310-linux_x86_64.whl#sha256=34ee585edf922a163b3ae6dbb90b291317947b7c8b47c045db058d92ec79e2b0
pytz==2024.2
PyYAML==6.0.2
pyzmq==26.2.0
requests==2.32.3
rich==13.9.4
scipy==1.14.1
simplejson==3.19.3
six==1.16.0
stack-data==0.6.3
sympy==1.13.1
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorflow==2.18.0
tensorflow-addons==0.23.0
tensorflow-hub==0.16.1
tensorflow-io-gcs-filesystem==0.37.1
tensorflow_decision_forests==1.11.0
tensorflowjs==4.22.0
tensorstore==0.1.68
termcolor==2.5.0
tf_keras==2.18.0
toolz==1.0.0
torch @ file:///home/abendafi/gan_model/torch-2.1.2%2Brocm6.1.3-cp310-cp310-linux_x86_64.whl#sha256=2cb85d429ce2108ba4b060297947ab09f86fac8c09468730d6db02a9221dc69e
torchmetrics==1.6.0
torchvision @ file:///home/abendafi/gan_model/torchvision-0.16.1%2Brocm6.1.3-cp310-cp310-linux_x86_64.whl#sha256=8949a908176563c7e310bcacbaba1fe2d6751a701051a85be976ca557ab435be
tornado==6.4.1
tqdm==4.67.0
traitlets==5.14.3
typeguard==2.13.3
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
wcwidth==0.2.13
Werkzeug==3.1.3
wrapt==1.17.0
wurlitzer==3.1.1
yarl==1.18.0
ydf==0.8.0
zipp==3.21.0
